Move isUgradeable to /src/lib
